refactor Address model to Score
